feat: Refactor profile feature and implement settings screen
This commit refactors the profile feature by moving screens to subdirectories and introduces a new settings screen.

- Moved AboutUsScreen, AccountDetailsScreen, ChangePasswordScreen, PrivacyPolicyScreen, ShippingAndReturnPolicyScreen, and TermsAndConditionsScreen to respective subdirectories under the profile feature.
- Updated imports in `lib/core/router/route.dart` to reflect the new file structure.
- Created a new `SettingScreen` with custom UI elements for managing app settings like language, theme, and notifications.
- Implemented `CustomAccountListTile` widget for consistent list tile styling in the profile section.
- Updated `ProfileScreen` to use the new `CustomAccountListTile` and structure the UI with containers.
- Created `customContainerView` widget for settings screen.
