+++ lib/feature/profile/setting/view/setting_screen.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+import 'package:mzaodina_app/core/resources/resources.dart';
+import 'package:mzaodina_app/feature/profile/setting/view/widget/custom_container_view.dart';
+import 'package:mzaodina_app/feature/profile/view/widget/custom_appbar_accounet.dart';
+
+class SettingScreen extends StatelessWidget {
+  const SettingScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: R.colors.whiteLight,
+      body: Padding(
+        padding: EdgeInsets.symmetric(horizontal: 16),
+        child: Column(
+          children: [
+            SafeArea(child: CustomAppBarAccount(title: 'الاعدادات')),
+            customContainerView(),
+          ],
+        ),
+      ),
+    );
+  }
+}
